refactor: Add context API
